pyproject.toml
[project]
name = "mcp-gateway"
version = "0.1.0"
description = "MCP Gateway - translate MCP tool-callings to HTTP requests"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.6.0",
    "pyyaml>=6.0.2",
]
